added code to run roberta model
